% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquid_gauge.R
\name{liquid_gauge}
\alias{liquid_gauge}
\title{D3 Liquid Fill Gauge}
\usage{
liquid_gauge(number, options = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
Generate D3 Liquid Fill Gauge
[url](http://bl.ocks.org/brattonc/5e5ce9beee483220e2f6)
}

